improve/fix v2.x.x integration
