+++ package.json
@@ -20,8 +20,8 @@
   },
   "homepage": "https://github.com/zondax/ledger-cosmos-js",
   "dependencies": {
-    "@babel/runtime": "^7.4.4",
-    "@ledgerhq/hw-transport": "^4.35.0",
+    "@babel/runtime": "^7.6.3",
+    "@ledgerhq/hw-transport": "^4.73.2",
     "bech32": "^1.1.3",
     "ripemd160": "^2.0.2"
   },
@@ -31,9 +31,9 @@
     "@babel/node": "^7.6.3",
     "@babel/plugin-transform-runtime": "^7.6.2",
     "@babel/preset-env": "^7.6.3",
-    "@ledgerhq/hw-transport-node-hid": "^4.72.1",
-    "@ledgerhq/hw-transport-u2f": "^4.72.1",
-    "@ledgerhq/hw-transport-webusb": "^4.72.1",
+    "@ledgerhq/hw-transport-node-hid": "^4.73.2",
+    "@ledgerhq/hw-transport-u2f": "^4.73.2",
+    "@ledgerhq/hw-transport-webusb": "^4.73.2",
     "@ledgerhq/logs": "^4.72.0",
     "@vue/cli-plugin-babel": "^3.12.0",
     "@vue/cli-plugin-eslint": "^3.12.0",
@@ -51,7 +51,9 @@
     "eslint-plugin-jest": "^22.19.0",
     "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-vue": "^5.2.3",
+    "index.js": "^0.0.3",
     "jest": "^24.9.0",
+    "jest-serial-runner": "^1.1.0",
     "prettier": "^1.18.2",
     "secp256k1": "^3.7.1",
     "vue": "^2.6.10",
