+++ src/helperV2.js
@@ -1,0 +1,29 @@
+import { signSendChunkv1 } from "./helperV1";
+import { PAYLOAD_TYPE } from "./common";
+
+export function serializePathv2(path) {
+  if (!path || path.length !== 5) {
+    throw new Error("Invalid path.");
+  }
+
+  const buf = Buffer.alloc(20);
+  buf.writeUInt32LE(0x80000000 + path[0], 0);
+  buf.writeUInt32LE(0x80000000 + path[1], 4);
+  buf.writeUInt32LE(0x80000000 + path[2], 8);
+  buf.writeUInt32LE(path[3], 12);
+  buf.writeUInt32LE(path[4], 16);
+
+  return buf;
+}
+
+export async function signSendChunkv2(app, chunkIdx, chunkNum, chunk) {
+  let payloadType = PAYLOAD_TYPE.ADD;
+  if (chunkIdx === 1) {
+    payloadType = PAYLOAD_TYPE.INIT;
+  }
+  if (chunkIdx === chunkNum) {
+    payloadType = PAYLOAD_TYPE.LAST;
+  }
+
+  return signSendChunkv1(app, payloadType, 0, chunk);
+}
