+++ tests/jest.js
@@ -1,0 +1,2 @@
+export default jest;
+export const { expect, test } = global;
