+++ vue_example/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="app">
     <img alt="Zondax logo" src="./assets/logo.png" width="300" />
-    <h1>Kusama integration example</h1>
+    <h1>Cosmos integration example</h1>
     <LedgerExample msg="Minimal integration example" />
   </div>
 </template>
@@ -12,8 +12,8 @@ import LedgerExample from "./components/LedgerExample.vue";
 export default {
   name: "App",
   components: {
-    LedgerExample
-  }
+    LedgerExample,
+  },
 };
 </script>
